Add commands and wrappers for working with the TPM 1.2 emulator.
The emulator needs to be started up, and the CreateEndorsementKeyPair
operation needs to be called on it as part of taking ownership.

The wrapper in run_other.go supports the mode of operation of the unix
socket exposed by the TPM 1.2 emulator.

This also adds a short doc that explains how to run tests using the IBM
software TPM 1.2 emulator.